requirements.txt
alabaster==1.0.0
anaconda==0.0.1.1
anyio==4.4.0
anytree==2.12.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==23.2.0
Automat==24.8.1
autopage==0.5.2
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
Brotli==1.1.0
brutils==2.2.0
bs4==0.0.2
bullet==2.2.0
cachetools==5.4.0
certifi==2022.12.7
cffi==1.17.0
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
click-plugins==1.1.1
cliff==4.7.0
cloudscraper==1.2.71
cmd2==2.4.3
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.7
dnspython==2.2.1
docutils==0.21.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.55.3
fqdn==1.5.1
future==1.0.0
gevent==24.11.1
geventhttpclient==2.3.3
gnupg==2.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
icmplib==3.0.4
idna==3.7
imagesize==1.4.1
impacket==0.10.0
incremental==24.7.2
ipwhois==1.2.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
ldap3==2.9.1
ldapdomaindump==0.9.4
locust==2.32.9
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
maxminddb==2.6.3
mccabe==0.7.0
mistune==3.0.2
msgpack==1.1.0
multidict==6.1.0
mysql-connector==2.2.9
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.4
ordered-set==4.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
phonenumbers==8.13.43
pillow==11.0.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
property-manager==3.0
Protego==0.3.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.21.0
PyDispatcher==2.0.7
pyflakes==3.2.0
Pygments==2.18.0
PyLaTeX==1.4.2
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyperclip==1.9.0
pypng==0.20220715.0
pyproject-api==1.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyRoxy @ git+https://github.com/MatrixTM/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
PySocks==1.7.1
pyswip==0.2.10
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-nmap==0.7.1
python-secrets==23.4.2
python-telegram-bot==20.5
python3-nmap==1.6.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
queuelib==1.7.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
Scrapy==2.11.2
seaborn==0.13.2
selenium==4.22.0
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==73.0.1
setuptools-scm==8.1.0
shodan==1.31.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
stevedore==5.2.0
style==1.1.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.1.2
tornado==6.4.2
tox==4.18.0
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.7.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
update==0.0.1
uri-template==1.3.0
urllib3==1.26.20
verboselogs==1.7
virtualenv==20.26.3
w3lib==2.2.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xkcdpass==1.19.9
XlsxWriter==3.2.0
yamlreader==3.0.4
yarl==1.18.3
zope.event==5.0
zope.interface==7.0.2
